added services, edited securityconfig

+++ src/main/java/com/example/IPRWC/Backend/controllers/ProductController.java
@@ -1,9 +1,12 @@
 package com.example.IPRWC.Backend.controllers;
 
 import com.example.IPRWC.Backend.entities.Product;
-import org.springframework.web.bind.annotation.*;
+import com.example.IPRWC.Backend.payload.response.MessageResponse;
 import com.example.IPRWC.Backend.repository.ProductRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 import java.util.Optional;
@@ -17,22 +20,28 @@ public class ProductController {
     ProductRepository productRepository;
 
     @GetMapping("/getAll")
-    public List<Product> getAllProducts() {
-        return productRepository.findAll();
+    public ResponseEntity<?> getAllProducts() {
+        if (!productRepository.findAll().isEmpty())
+            return new ResponseEntity<>(productRepository.findAll(), HttpStatus.OK);
+        return new ResponseEntity<>("Products not found", HttpStatus.NOT_FOUND);
     }
 
     @GetMapping("/get/{id}")
-    public Optional<Product> getProduct(@PathVariable Long id) {
-        return productRepository.findById(id);
+    public ResponseEntity<?> getProduct(@PathVariable Long id) {
+        if (productRepository.findById(id).isPresent()) {
+            return new ResponseEntity<>(productRepository.findById(id), HttpStatus.OK);
+        }
+        return new ResponseEntity<>("Product with given id not found", HttpStatus.NOT_FOUND);
     }
 
     @PostMapping("/add")
-    public Product saveProduct(@RequestBody Product product) {
-        return productRepository.save(product);
+    public ResponseEntity<?> saveProduct(@RequestBody Product product) {
+        productRepository.save(product);
+        return ResponseEntity.ok(new MessageResponse("Product added!"));
     }
 
     @PutMapping("/edit/{id}")
-    public Product putProduct(@RequestBody Product product, @PathVariable Long id) {
+    public ResponseEntity<?> putProduct(@RequestBody Product product, @PathVariable Long id) {
         Optional<Product> productData = productRepository.findById(id);
         if (productData.isPresent()) {
             Product _product = productData.get();
@@ -40,14 +49,19 @@ public class ProductController {
             _product.setProduct_rating(product.getProduct_rating());
             _product.setDescription(product.getDescription());
             _product.setProduct_name(product.getProduct_name());
-            return productRepository.save(_product);
+            productRepository.save(_product);
+            return ResponseEntity.ok(new MessageResponse("Product with id: " + _product.getId() + " has been edited"));
         } else {
-            return null;
+            return ResponseEntity.badRequest().body(new MessageResponse("Product not edited"));
         }
     }
 
     @DeleteMapping("/delete/{id}")
-    public void deleteProduct(@PathVariable Long id) {
-        productRepository.deleteById(id);
+    public ResponseEntity<?> deleteProduct(@PathVariable Long id) {
+        if (productRepository.findById(id).isPresent()) {
+            productRepository.deleteById(id);
+            return ResponseEntity.ok().body(new MessageResponse("Product deleted"));
+        }
+        return ResponseEntity.unprocessableEntity().body("Product doesn't exist!");
     }
 }
