+++ src/main/java/com/example/IPRWC/Backend/controllers/UserController.java
@@ -3,6 +3,8 @@ package com.example.IPRWC.Backend.controllers;
 import com.example.IPRWC.Backend.entities.User;
 import com.example.IPRWC.Backend.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
@@ -25,17 +27,20 @@ public class UserController {
 
 
     @GetMapping("/all")
-    public List<User> getAllUsers(){
+    public List<User> getAllUsers() {
         return userRepository.findAll();
     }
 
     @GetMapping("/get/{id}")
-    public Optional<User> getUser(@PathVariable Long id){
-        return userRepository.findById(id);
+    public ResponseEntity<?> getUser(@PathVariable Long id) {
+        if (userRepository.findById(id).isPresent()) {
+            return new ResponseEntity<>(userRepository.findById(id), HttpStatus.OK);
+        }
+        return new ResponseEntity<>("User with ID not found" + id, HttpStatus.NOT_FOUND);
     }
 
     @PutMapping("/edit")
-    public User editUserDetails(@RequestBody User user) {
+    public ResponseEntity<?> editUserDetails(@RequestBody User user) {
         String email = SecurityContextHolder.getContext().getAuthentication().getName();
         Optional<User> userData = userRepository.findByEmail(email);
         if (userData.isPresent() && SecurityContextHolder.getContext().getAuthentication().isAuthenticated()) {
@@ -49,14 +54,17 @@ public class UserController {
             _user.setPhoneNumber(user.getPhoneNumber());
             _user.setPlace(user.getPlace());
 
-            return userRepository.save(_user);
+            userRepository.save(_user);
+            return new ResponseEntity<>("User Edited!", HttpStatus.OK);
+        } else if (userData.isEmpty()) {
+            return new ResponseEntity<>("User doesn't exist", HttpStatus.NOT_FOUND);
         }
-        return null;
+        return new ResponseEntity<>("Something went wrong!", HttpStatus.UNPROCESSABLE_ENTITY);
     }
 
     @PreAuthorize("hasAuthority('ROLE_ADMIN')")
     @PutMapping("edit/{id}")
-    public User editUserDetails(@PathVariable Long id,@RequestBody User user) {
+    public ResponseEntity<?> editUserDetails(@PathVariable Long id, @RequestBody User user) {
         Optional<User> userData = userRepository.findById(id);
         if (userData.isPresent()) {
             System.out.println("Hij werkt!");
@@ -70,14 +78,21 @@ public class UserController {
             _user.setPhoneNumber(user.getPhoneNumber());
             _user.setPlace(user.getPlace());
 
-            return userRepository.save(_user);
+            userRepository.save(_user);
+            return new ResponseEntity<>("User edited", HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>("User not found", HttpStatus.NOT_FOUND);
         }
-        return null;
     }
 
     @PreAuthorize("hasAuthority('ROLE_ADMIN')")
     @DeleteMapping("/delete/{id}")
-    public void deleteUser(@PathVariable Long id){
-        userRepository.deleteById(id);
+    public ResponseEntity<?> deleteUser(@PathVariable Long id) {
+        if (userRepository.findById(id).isPresent()){
+            userRepository.deleteById(id);
+            return new ResponseEntity<>("User deleted!", HttpStatus.OK);
+        }
+        return new ResponseEntity<>("User doesn't exist!", HttpStatus.NOT_FOUND);
+
     }
 }
